requirements.txt
# Core
fastapi==0.95.0
pydantic==1.10.7
python-dotenv==1.0.0
aiohttp>=3.9.0  # Instead of aiohttp==3.8.4

# MongoDB
motor==3.1.2
pymongo==4.3.3

# Web scraping
beautifulsoup4==4.12.0
httpx==0.23.3

# Templating and document generation
jinja2==3.1.2
pdfkit==1.0.0
python-docx==0.8.11
fpdf2==2.7.4

# Utilities
markdownify==0.11.6
python-dateutil==2.8.2
tqdm==4.65.0